Deletes homescreen shortcut
When a note is created and a shortcut for it is made on the home screen. When the note itself is deleted in the app, the shortcut does not automatically disappear, instead an error occurs when the shortcut is clicked. This added method ("removershortCut") properly deletes the shortcut on the home screen whenever the note itself is deleted.